diagnostic: port to generic layer.
